SAMZA-1602: Moving class StreamAssert from src/test to src/main
Tested with ./gradlew clean build, which is passing

Author: sanil15 <sanil.jain15@gmail.com>

Reviewers: Xinyu Liu <xinyuliu.us@gmail.com>

Closes #439 from Sanil15/SAMZA-1602
